Decouple delegates into concrete classes
