+++ jvm-testing/src/main/kotlin/app/cash/trifle/testing/TestEndEntity.kt
@@ -3,14 +3,14 @@ package app.cash.trifle.testing
 import app.cash.trifle.Certificate
 import app.cash.trifle.CertificateRequest
 import app.cash.trifle.SignedData
-import app.cash.trifle.Trifle
+import app.cash.trifle.delegates.EndEntity
 import app.cash.trifle.extensions.CertificateChain
 
 /**
  * Trifle End Entity used strictly for testing.
  */
 data class TestEndEntity internal constructor(
-  private val endEntity: Trifle.EndEntity,
+  private val endEntity: EndEntity,
   val certChain: CertificateChain,
   val certRequest: CertificateRequest,
   val certificate: Certificate = certChain.first()
